Fix graceful shutdown test
It was failing, at least I thought it was. I think I only thought it
was, because it logged an exception, but the test didn't actually fail
when I re-ran it today. This resorts to using a future, that we block
wait for after the shutdown has been triggered. This was I can get the
exception thrown in the same thread as JUnit, and also more properly
format the test as when, then, assert.

I also took the time to rewrite the test a bit to fall somewhere
between my natural style and @tipsy's preferred style.