Setup Compression Testing
Testing compression with JBrotli. Primarily looking for differences between Compressor and StreamCompressor.

Initial results are promising, with performance comparable and even surpassing that of Javalin's gzip implementation.

It seems the crashes I experienced on my previous test runs are caused specifically by using the BrotliOutputStream object, which uses the StreamCompressor with a byte buffer. Manually invoking BrotliStreamCompressor.compressArray() and feeding it a byte array does successfully avoid the JVM crash, but only encodes the first few bytes of the array. This is due to a documented bug with older versions of Brotli that are still being used by this library. Here is a link to the bug: https://github.com/google/brotli/issues/346

As a result, I was only able to achieve successful compression by using the Compressor class (not StreamCompressor).

I may do some research into updating JBrotli to use the newest version of the Brotli lib, though right now I do not know how much work this would entail.