Updated Wrapper Code
Updated code in Brotli wrapper. The current version (with Brotli compression level 4) is just as fast and memory efficient as Gzip while offering superior compression capabilities.

At least this is according to a simple test with the same object structure repeated many times over.

There are some articles about this on the internet that support my findings on a more varied and comprehensive set of benchmarks, however. So unless proven otherwise, I reasonably believe this to be true