Static Brotli Implementation
The way I am doing static Brotli compression, is by inheriting from Jetty's GzipHandler and overriding its handle() method. This makes sense because all the "operational" logic is found in the overriden method, meaning that if Brotli compression cannot be performed for whatever reason (unsupported, disabled, etc.), I can literally just call the original handle method via super()

I also implemented a new interceptor class called BrotliHttpOutputInterceptor. This is directly based off (but not inherited from) the corresponding GzipHttpOutputInterceptor and effectively does the same form of handling, just specific to Brotli instead of Gzip.

This is the best approach I could think of, as it falls exactly in line with how Jetty's built-in GzipHandler does its magic. It checks that conditions for gzip are matched, then sets an interceptor on baseRequest.getResponse() for further handling/modification to the response before it is sent back to client.