Bugfix for 671
The bug was first discovered when including (and responding with) large webjars (e.g. Buefy). It was caused at line 688 in ResourceService.java, where jetty writes content via a buffer method rather than the usual single-pass write on line 686.

This resulted in multiple calls to our overriden write method, and as a result, each chunk of data was being compressed separately.

To get around the problem, the OutputStreamWrapper class has been changed to write to a dummy object rather than the actual outputStream. This object (ByteArrayOutputStream) collects and aggregates everything that jetty and/or javalin try writing to output.

Once all response handling by javalin/jetty is finished, we call the new finalizeWrite() method, which takes the data from the dummy object, processes it as required (compression), then copies it to the real output stream.