Cleaner Fix
We check if a content_encoding header already exists on first write. If it does, we never enable the compressors at all.

This way, we can rely on compression enable vars (gzipEnabled and brotliEnabled) to tell us whether or not we actually did any compression work.

If not, we just skip the steps in the finalize() code.