Add Redoc Options configuration
Redoc contains a list of customizable options in its options object (https://github.com/Redocly/redoc#redoc-options-object). This commit enables that customization inside Javalin's DSL.

For each type of object, a builder was created to ease Java's usage, as the language doesn't support default arguments yet.

An important point to verify here is that the Theme option, although configurable, won't work with the standalone version currently, due to a bug in Redoc itself (https://github.com/Redocly/redoc/issues/1248). If it gets fixed, the code from this commit will automatically work with it.

Closes #1097

Closes #