Allow comments to be sent in sseClient
The sseClient spec allows comment blocks to be sent, which can be useful when implementing a keepalive.

To guard against sending 'other' data, ensured multiline comments passed in should be split and all sent as comment.

I really don't know kotlin at all, so happy to take feedback on what I've done and clean it up...