[websocket] Fix bug in disableAutomaticPings not disabling pings
This patch moves the ping future out of the WsContext. As it turns out,
the WsContext seems to be reinitialized often, dropping the reference
to the pingFuture, i.e. setting it to null. This made it impossible to
disable automatic pings.

This patch moves all ping futures to a separate ConcurrentHashMap with
the context as a key.