Make a single decision to compress or not
Use a boolean to more clearly guard the evaluation of whether to
compress or not, ensuring that the decision is made only once. A null
assignment to `compressedStream` was previously used to guard this
evaluation, but a null is a legitimate outcome of a compression choice,
where we don't want to reevaluate our choice again.

The `write(byte: Int)` function uses the compression stream now too.

A request scope attribute `ALWAYS_COMPRESS` signals when compression
should always be used (when allowed). Because the intent of function
`writeJsonStream(stream: Stream<*>)` is to write a lot of data, and the
total size of data cannot be known, it is reasonable to set the
`ALWAYS_COMPRESS` attribute for that function call.