Draft for demonstrating reflection reduction via `NoClassDefFoundError`
This is a hack, but I've done it a lot and personally think it could be
a better way of handling optional dependencies over `Class.forName`

The thing is, `Class.forName` was designed for this very use-case of
optional dependencies. Javalin claims "no reflection" even in the very
readme file, so there are two ways to back this claim

- Remove all non-test usage of reflection entirely, opting for
  techniques like these (which work and in my opinion have no downsides,
  except that readers might be annoyed we are catching `Error`s?

  This even works properly in a `native-image` as well)

- Remove the claim of "no reflection"

Please note that I have no clue on writing idiomatic Kotlin.

Related-to: https://github.com/avaje/avaje-http/pull/670#issuecomment-3567115136
Related-to: https://github.com/avaje/avaje-inject/pull/932
Related-to: https://github.com/avaje/avaje-http/pull/675
Signed-off-by: Mahied Maruf <contact@mechite.com>